Updated Coolant Controls
Coolant control now has Mist and Flood

+++ spindle_control.h
@@ -24,7 +24,7 @@
 #include <avr/io.h>
 
 void spindle_init();
-void spindle_run(int direction, uint32_t rpm);
+void spindle_run(int8_t direction, uint16_t rpm);
 void spindle_stop();
 
 #endif
